requirements.txt
Eve==0.7.4
Flask==0.12
Flask-Cors==3.0.3
pymongo==3.4.0
pyjwt==1.5.2
